refactor(auth): reorganize auth utilities and inline wallet creation
- Move permissions.ts to utils/permissions.ts for better organization
- Remove standalone wallet.ts helper file
- Inline wallet creation logic directly in create-wallet mutation
- Add local development wallet funding to create-wallet mutation
- Update import paths to reflect new utils structure
- Improve error handling in wallet creation with proper Portal client usage

This refactoring consolidates auth-related utilities and removes unnecessary
indirection while maintaining the same functionality. The wallet creation
logic is now co-located with the mutation that uses it, making the code
more maintainable and easier to understand.
