+++ kit/dapp/src/lib/auth/utils/permissions.ts
