+++ kit/dapp/src/orpc/middlewares/auth/offchain-permissions.middleware.ts
@@ -1,5 +1,5 @@
 import { authClient } from "@/lib/auth/auth.client";
-import type { Permissions } from "@/lib/auth/permissions";
+import type { Permissions } from "@/lib/auth/utils/permissions";
 import type { Context } from "@/orpc/context/context";
 import { baseRouter } from "@/orpc/procedures/base.router";
 import { z } from "zod";
